:hammer: Move plot utils
