+++ torch_uncertainty/utils/__init__.py
@@ -2,4 +2,4 @@
 from .checkpoints import get_version
 from .cli import TULightningCLI
 from .hub import load_hf
-from .misc import csv_writer
+from .misc import csv_writer, plot_hist
